try to fix std::min call on all compilers

+++ src/mongo/s/d_migrate.cpp
@@ -421,7 +421,7 @@ namespace mongo {
             if ( totalRecs > 0 ) {
                 avgRecSize = d->stats.datasize / totalRecs;
                 maxRecsWhenFull = maxChunkSize / avgRecSize;
-                maxRecsWhenFull = std::min( Chunk::MaxObjectPerChunk + 1 , 130 * maxRecsWhenFull / 100 /* slack */ );
+                maxRecsWhenFull = std::min( (unsigned long long)(Chunk::MaxObjectPerChunk + 1) , 130 * maxRecsWhenFull / 100 /* slack */ );
             }
             else {
                 avgRecSize = 0;
